Clean up ExtraRandomSplitterTest.
